Add navi command to NaviPoint, update navi commands in ATM and Job, add nearest navi point retrieving in NavigationService, update NearestNaviPointCommand
