+++ core/src/main/java/com/rettichlp/unicacityaddon/base/enums/location/Job.java
@@ -40,7 +40,7 @@ public enum Job {
     private final String command;
 
     public String getNaviCommand() {
-        return "/navi " + getX() + "/" + getY() + "/" + getZ();
+        return "/navi " + x + "/" + y + "/" + z;
     }
 
     public FloatVector3 getLocation() {
